ml_prod_deploy/authentication_token/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
DateTime==4.9
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==1.4.45
threadpoolctl==3.1.0
urllib3==1.26.13
uuid==1.30
Werkzeug==2.2.2
zope.interface==5.5.2
